make preemption optional and in a separate class
closes #7
