+++ config.js
@@ -1,13 +1,16 @@
 module.exports = {
-  preeemptionPeriod: 10,
-  totalCapacity: 5,
-  preemption_high_percent: 0.75,
-  preemption_low_percent: 0.25,
-  factors: {
-    "nodejs": 1,
-    "nodejs:6": 1,
-    "java": 3,
-    "blackbox": 5
+  total_capacity: 5,
+  preemption: {
+    enabled: false,
+    period: 10,
+    high_percent: 0.75,
+    low_percent: 0.25,
+    factors: {
+      "nodejs": 1,
+      "nodejs:6": 1,
+      "java": 3,
+      "blackbox": 5
+    },
   },
   retries: {
 	  timeout: 5000, //msec
